Move everything around for COPR.
COPR's SCM builds don't do the right thing if there's anything other
than a spec file and a README in a directory, so move every spec file to
its own directory. Add a bunch of READMEs with status icons and
dependencies and build orders and whatnot.

Remove makedeps.sh since this isn't being built in a local mock anymore.